# security/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Security Example API","description":"This API demonstrates the use of the goa security DSL","version":""},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/another/secure":{"get":{"tags":["another_secured_service"],"summary":"secure another_secured_service","description":"This action is secured with the jwt scheme\n\n**Required security scopes for jwt**:\n  * `api:read`","operationId":"another_secured_service#secure","parameters":[{"name":"fail","in":"query","description":"Whether to force auth failure even with a valid JWT","required":false,"type":"boolean"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceSecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceSecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]},"put":{"tags":["another_secured_service"],"summary":"doubly_secure another_secured_service","description":"This action is secured with the jwt scheme and also requires an API key query string.\n\n**Required security scopes for jwt**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for api_key**:\n  * `api:read`\n  * `api:write`","operationId":"another_secured_service#doubly_secure","parameters":[{"name":"k","in":"query","description":"API key","required":true,"type":"string"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceDoublySecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceDoublySecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"api_key_query_k":[],"jwt_header_Authorization":[]}]},"post":{"tags":["another_secured_service"],"summary":"also_doubly_secure another_secured_service","description":"This action is secured with the jwt scheme and also requires an API key header.\n\n**Required security scopes for jwt**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for api_key**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for basic**:\n  * `api:read`\n  * `api:write`","operationId":"another_secured_service#also_doubly_secure","parameters":[{"name":"k","in":"query","description":"API key","required":false,"type":"string"},{"name":"oauth","in":"query","required":false,"type":"string"},{"name":"X-Authorization","in":"header","description":"JWT used for authentication","required":false,"type":"string"},{"name":"Authorization","in":"header","description":"Basic Auth security using Basic scheme (https://tools.ietf.org/html/rfc7617)","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceAlsoDoublySecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceAlsoDoublySecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"api_key_query_k":[],"jwt_header_X-Authorization":[]},{"basic_header_Authorization":[],"oauth2_query_oauth":["api:read","api:write"]}]}},"/another/signin":{"post":{"tags":["another_secured_service"],"summary":"signin another_secured_service","description":"Creates a valid JWT","operationId":"another_secured_service#signin","parameters":[{"name":"Authorization","in":"header","description":"Basic Auth security using Basic scheme (https://tools.ietf.org/html/rfc7617)","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceSigninResponseBody","required":["jwt","api_key","oauth_token"]}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/AnotherSecuredServiceSigninUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"basic_header_Authorization":[]}]}},"/secure":{"get":{"tags":["secured_service"],"summary":"secure secured_service","description":"This action is secured with the jwt scheme\n\n**Required security scopes for jwt**:\n  * `api:read`","operationId":"secured_service#secure","parameters":[{"name":"fail","in":"query","description":"Whether to force auth failure even with a valid JWT","required":false,"type":"boolean"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/SecuredServiceSecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/SecuredServiceSecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]},"put":{"tags":["secured_service"],"summary":"doubly_secure secured_service","description":"This action is secured with the jwt scheme and also requires an API key query string.\n\n**Required security scopes for jwt**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for api_key**:\n  * `api:read`\n  * `api:write`","operationId":"secured_service#doubly_secure","parameters":[{"name":"k","in":"query","description":"API key","required":true,"type":"string"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/SecuredServiceDoublySecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/SecuredServiceDoublySecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"api_key_query_k":[],"jwt_header_Authorization":[]}]},"post":{"tags":["secured_service"],"summary":"also_doubly_secure secured_service","description":"This action is secured with the jwt scheme and also requires an API key header.\n\n**Required security scopes for jwt**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for api_key**:\n  * `api:read`\n  * `api:write`\n\n**Required security scopes for basic**:\n  * `api:read`\n  * `api:write`","operationId":"secured_service#also_doubly_secure","parameters":[{"name":"k","in":"query","description":"API key","required":false,"type":"string"},{"name":"oauth","in":"query","required":false,"type":"string"},{"name":"X-Authorization","in":"header","description":"JWT used for authentication","required":false,"type":"string"},{"name":"Authorization","in":"header","description":"Basic Auth security using Basic scheme (https://tools.ietf.org/html/rfc7617)","required":false,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"string"}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/SecuredServiceAlsoDoublySecureUnauthorizedResponseBody"}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/SecuredServiceAlsoDoublySecureInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"api_key_query_k":[],"jwt_header_X-Authorization":[]},{"basic_header_Authorization":[],"oauth2_query_oauth":["api:read","api:write"]}]}},"/signin":{"post":{"tags":["secured_service"],"summary":"signin secured_service","description":"Creates a valid JWT","operationId":"secured_service#signin","parameters":[{"name":"Authorization","in":"header","description":"Basic Auth security using Basic scheme (https://tools.ietf.org/html/rfc7617)","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/SecuredServiceSigninResponseBody","required":["jwt","api_key","oauth_token"]}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/SecuredServiceSigninUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"basic_header_Authorization":[]}]}}},"definitions":{"AnotherSecuredServiceAlsoDoublySecureInvalidScopesResponseBody":{"title":"AnotherSecuredServiceAlsoDoublySecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Nam et consequuntur nemo sit cum."},"AnotherSecuredServiceAlsoDoublySecureUnauthorizedResponseBody":{"title":"AnotherSecuredServiceAlsoDoublySecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Molestiae quo."},"AnotherSecuredServiceDoublySecureInvalidScopesResponseBody":{"title":"AnotherSecuredServiceDoublySecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Ad explicabo sed."},"AnotherSecuredServiceDoublySecureUnauthorizedResponseBody":{"title":"AnotherSecuredServiceDoublySecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Dolorem et totam nemo numquam eum."},"AnotherSecuredServiceSecureInvalidScopesResponseBody":{"title":"AnotherSecuredServiceSecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Eaque pariatur eos laborum."},"AnotherSecuredServiceSecureUnauthorizedResponseBody":{"title":"AnotherSecuredServiceSecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Consequatur blanditiis consectetur quis."},"AnotherSecuredServiceSigninResponseBody":{"title":"AnotherSecuredServiceSigninResponseBody","type":"object","properties":{"api_key":{"type":"string","description":"API Key","example":"abcdef12345"},"jwt":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"oauth_token":{"type":"string","description":"OAuth2 token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"example":{"api_key":"abcdef12345","jwt":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ","oauth_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"required":["jwt","api_key","oauth_token"]},"AnotherSecuredServiceSigninUnauthorizedResponseBody":{"title":"AnotherSecuredServiceSigninUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Ipsum modi saepe."},"SecuredServiceAlsoDoublySecureInvalidScopesResponseBody":{"title":"SecuredServiceAlsoDoublySecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Error et et officiis unde."},"SecuredServiceAlsoDoublySecureUnauthorizedResponseBody":{"title":"SecuredServiceAlsoDoublySecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Alias architecto quasi."},"SecuredServiceDoublySecureInvalidScopesResponseBody":{"title":"SecuredServiceDoublySecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Error dolorem deserunt similique ex."},"SecuredServiceDoublySecureUnauthorizedResponseBody":{"title":"SecuredServiceDoublySecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Delectus ratione voluptas sit quo."},"SecuredServiceSecureInvalidScopesResponseBody":{"title":"SecuredServiceSecureInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Sunt id facere est suscipit provident veritatis."},"SecuredServiceSecureUnauthorizedResponseBody":{"title":"SecuredServiceSecureUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Nesciunt molestiae consequuntur libero."},"SecuredServiceSigninResponseBody":{"title":"SecuredServiceSigninResponseBody","type":"object","properties":{"api_key":{"type":"string","description":"API Key","example":"abcdef12345"},"jwt":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"oauth_token":{"type":"string","description":"OAuth2 token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"example":{"api_key":"abcdef12345","jwt":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ","oauth_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"required":["jwt","api_key","oauth_token"]},"SecuredServiceSigninUnauthorizedResponseBody":{"title":"SecuredServiceSigninUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Amet voluptas quia."}},"securityDefinitions":{"api_key_query_k":{"type":"apiKey","description":"Secures endpoint by requiring an API key.","name":"k","in":"query"},"basic_header_Authorization":{"type":"basic","description":"Basic authentication used to authenticate security principal during signin"},"jwt_header_Authorization":{"type":"apiKey","description":"Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes \"api:read\" and \"api:write\".\n\n**Security Scopes**:\n  * `api:read`: Read-only access\n  * `api:write`: Read and write access","name":"Authorization","in":"header"},"jwt_header_X-Authorization":{"type":"apiKey","description":"Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes \"api:read\" and \"api:write\".\n\n**Security Scopes**:\n  * `api:read`: Read-only access\n  * `api:write`: Read and write access","name":"X-Authorization","in":"header"},"oauth2_query_oauth":{"type":"oauth2","description":"Secures endpoint by requiring a valid OAuth2 token retrieved via the signin endpoint. Supports scopes \"api:read\" and \"api:write\".","flow":"accessCode","authorizationUrl":"http://goa.design/authorization","tokenUrl":"http://goa.design/token","scopes":{"api:read":"Read-only access","api:write":"Read and write access"}}},"externalDocs":{"description":"Security example README","url":"https://github.com/goadesign/goa/tree/master/example/security/README.md"}}